+++ lib/rorvswild/agent.rb
@@ -159,10 +159,6 @@ module RorVsWild
       current_data[:error_context] = hash if current_data
     end
 
-    def send_server_timing=(boolean)
-      current_data[:send_server_timing] = boolean if current_data
-    end
-
     def current_data
       Thread.current[:rorvswild_data]
     end